refactor(cardano-services): nest ServiceNames in Http Server context, generalise utils and errors
