--- packages/cardano-services/src/Program/ServiceNames.ts
@@ -1,13 +1,0 @@
-/**
- * Used as mount segments, so must be URL-friendly
- *
- */
-export enum ServiceNames {
-  Asset = 'asset',
-  StakePool = 'stake-pool',
-  NetworkInfo = 'network-info',
-  TxSubmit = 'tx-submit',
-  Utxo = 'utxo',
-  ChainHistory = 'chain-history',
-  Rewards = 'rewards'
-}
