+++ packages/cardano-services/src/Program/errors/MissingProgramOption.ts
@@ -1,7 +1,6 @@
 import { CustomError } from 'ts-custom-error';
-import { ServiceNames } from '../ServiceNames';
 
-export class MissingProgramOption<OptionsType> extends CustomError {
+export class MissingProgramOption<ServiceNames, OptionsType> extends CustomError {
   public constructor(service: ServiceNames, option: OptionsType | OptionsType[]) {
     super();
     this.message = `${service} requires the ${Array.isArray(option) ? option.join(' or ') : option} program option.`;
