+++ packages/cardano-services/src/Program/index.ts
@@ -1,5 +1,4 @@
 export * from './errors';
 export * from './programs/';
-export * from './ServiceNames';
 export * from './utils';
 export * from './services';
