+++ packages/cardano-services/src/StakePool/StakePoolHttpService.ts
@@ -1,7 +1,7 @@
 import * as OpenApiValidator from 'express-openapi-validator';
 import { HttpService } from '../Http';
 import { Logger } from 'ts-log';
-import { ServiceNames } from '../Program/ServiceNames';
+import { ServiceNames } from '../Program';
 import { StakePoolProvider } from '@cardano-sdk/core';
 import { providerHandler } from '../util';
 import express from 'express';
